Create `files.py` which has no import issues.
