+++ plotly/tests/test_core/test_tools/test_configuration.py
@@ -2,14 +2,15 @@ from __future__ import absolute_import
 
 from unittest import TestCase
 
-from plotly.tools import get_config_defaults, _FILE_CONTENT, CONFIG_FILE
+from plotly.files import CONFIG_FILE, FILE_CONTENT
+from plotly.tools import get_config_defaults
 
 
 class TestGetConfigDefaults(TestCase):
 
     def test_config_dict_is_equivalent_copy(self):
 
-        original = _FILE_CONTENT[CONFIG_FILE]
+        original = FILE_CONTENT[CONFIG_FILE]
         copy = get_config_defaults()
         self.assertIsNot(copy, original)
         self.assertEqual(copy, original)
