+++ plotly/tools.py
@@ -20,6 +20,8 @@ import math
 from plotly import utils
 from plotly import exceptions
 from plotly import session
+from plotly.files import (CONFIG_FILE, CREDENTIALS_FILE, FILE_CONTENT,
+                          GRAPH_REFERENCE_FILE, check_file_permissions)
 
 
 # Warning format
@@ -48,39 +50,6 @@ try:
 except ImportError:
     _numpy_imported = False
 
-PLOTLY_DIR = os.path.join(os.path.expanduser("~"), ".plotly")
-CREDENTIALS_FILE = os.path.join(PLOTLY_DIR, ".credentials")
-CONFIG_FILE = os.path.join(PLOTLY_DIR, ".config")
-TEST_DIR = os.path.join(os.path.expanduser("~"), ".test")
-TEST_FILE = os.path.join(PLOTLY_DIR, ".permission_test")
-
-# this sets both the DEFAULTS and the TYPES for these items
-_FILE_CONTENT = {CREDENTIALS_FILE: {'username': '',
-                                    'api_key': '',
-                                    'proxy_username': '',
-                                    'proxy_password': '',
-                                    'stream_ids': []},
-                 CONFIG_FILE: {'plotly_domain': 'https://plot.ly',
-                               'plotly_streaming_domain': 'stream.plot.ly',
-                               'plotly_api_domain': 'https://api.plot.ly',
-                               'plotly_ssl_verification': True,
-                               'plotly_proxy_authorization': False,
-                               'world_readable': True}}
-
-
-try:
-    os.mkdir(TEST_DIR)
-    os.rmdir(TEST_DIR)
-    if not os.path.exists(PLOTLY_DIR):
-        os.mkdir(PLOTLY_DIR)
-    f = open(TEST_FILE, 'w')
-    f.write('testing\n')
-    f.close()
-    os.remove(TEST_FILE)
-    _file_permissions = True
-except:
-    _file_permissions = False
-
 
 def get_config_defaults():
     """
@@ -92,28 +61,27 @@ def get_config_defaults():
             # do something
 
     """
-    return dict(_FILE_CONTENT[CONFIG_FILE])  # performs a shallow copy
+    return dict(FILE_CONTENT[CONFIG_FILE])  # performs a shallow copy
 
 
-def check_file_permissions():
-    return _file_permissions
 
 
 def ensure_local_plotly_files():
     """Ensure that filesystem is setup/filled out in a valid way"""
-    if _file_permissions:
+    if check_file_permissions():
         for fn in [CREDENTIALS_FILE, CONFIG_FILE]:
             utils.ensure_file_exists(fn)
             contents = utils.load_json_dict(fn)
-            for key, val in list(_FILE_CONTENT[fn].items()):
+            for key, val in list(FILE_CONTENT[fn].items()):
                 # TODO: removed type checking below, may want to revisit
                 if key not in contents:
                     contents[key] = val
             contents_keys = list(contents.keys())
             for key in contents_keys:
-                if key not in _FILE_CONTENT[fn]:
+                if key not in FILE_CONTENT[fn]:
                     del contents[key]
             utils.save_json_dict(fn, contents)
+        ensure_graph_reference_file()
     else:
         warnings.warn("Looks like you don't have 'read-write' permission to "
                       "your 'home' ('~') directory or to our '~/.plotly' "
@@ -140,7 +108,7 @@ def set_credentials_file(username=None,
     :param (str) proxy_password: The pw associated with your Proxy un
 
     """
-    if not _file_permissions:
+    if not check_file_permissions():
         raise exceptions.PlotlyError("You don't have proper file permissions "
                                      "to run this function.")
     ensure_local_plotly_files()  # make sure what's there is OK
@@ -168,11 +136,11 @@ def get_credentials_file(*args):
         get_credentials_file('username')
 
     """
-    if _file_permissions:
+    if check_file_permissions():
         ensure_local_plotly_files()  # make sure what's there is OK
         return utils.load_json_dict(CREDENTIALS_FILE, *args)
     else:
-        return _FILE_CONTENT[CREDENTIALS_FILE]
+        return FILE_CONTENT[CREDENTIALS_FILE]
 
 
 def reset_credentials_file():
@@ -199,7 +167,7 @@ def set_config_file(plotly_domain=None,
     :param (bool) world_readable: True = public, False = private
 
     """
-    if not _file_permissions:
+    if not check_file_permissions():
         raise exceptions.PlotlyError("You don't have proper file permissions "
                                      "to run this function.")
     ensure_local_plotly_files()  # make sure what's there is OK
@@ -243,11 +211,11 @@ def get_config_file(*args):
         get_config_file('plotly_domain')
 
     """
-    if _file_permissions:
+    if check_file_permissions():
         ensure_local_plotly_files()  # make sure what's there is OK
         return utils.load_json_dict(CONFIG_FILE, *args)
     else:
-        return _FILE_CONTENT[CONFIG_FILE]
+        return FILE_CONTENT[CONFIG_FILE]
 
 
 def reset_config_file():
